requirements.txt
aiogram>=3.20.0
SQLAlchemy>=2.0.0
asyncpg>=0.29.0
redis>=6.0.0
alembic>=1.13.0
python-dotenv>=1.0.0 